Fix Gene object printing and type mismatches
- Fix Gene object string conversion: properly extract symbol index from Key
  by casting to Value first, then extracting payload with PAYLOAD_MASK
- Fix call base stack type mismatch: change uint16 to uint32 to match
  pop_call_base() return type
- Update examples/full.gene: add Gene object examples with correct symbol
  syntax (:Type instead of bare Type identifiers)
- Simplify string implementation: remove short string optimization
- Revert string append to in-place mutation for performance

+++ src/gene/vm.nim
@@ -484,7 +484,7 @@ proc print_instruction_profile*(self: VirtualMachine) =
 
 #################### Unified Callable System ####################
 
-proc pop_call_base_info(vm: VirtualMachine, expected: int = -1): tuple[hasBase: bool, base: uint16, count: int] {.inline.} =
+proc pop_call_base_info(vm: VirtualMachine, expected: int = -1): tuple[hasBase: bool, base: uint32, count: int] {.inline.} =
   ## Retrieve call base metadata if present, otherwise fall back to expected count.
   if vm.frame.call_bases.is_empty():
     result.hasBase = false
